Rewrite f32.xml to match dropped variant background files
As Luya wrote to the mailing list, "new change is the drop of
standard, normalish, tv, tv-wider folders in favour of a single
wallpaper" - but the GNOME XML file was not updated for this
change at all, it's still trying to point to the old files that
don't exist in the package any more. The format is not documented
anywhere I can find, but I think this should be correct to simply
point at that one single file, I don't know if any other magic is
necessary to tell GNOME how to scale it to different aspect
ratios or anything.

This also re-indents the file with indentation that isn't a mix
of tabs and spaces, and is actually correct.

Signed-off-by: Adam Williamson <awilliam@redhat.com>