Rename f36-time.xml back to f36.xml
It is not safe to just rename this file. The schema override in
desktop-backgrounds-gnome points to exactly
/usr/share/backgrounds/fNN/default/fNN.xml (where NN is the
release number); if that file does not exist, then new user
accounts will have a broken config. Also I think when the user
account is created this string is written into their config -
so even if we changed the override to point to this new name,
anyone who installed F36 already would still have a config that
looks for f36.xml and if there is no f36.xml their background
would break.

I'm not clear whether we want the animated background or the
static night one to be the default, but whichever it is should
be installed as f36.xml, not any other name. This assumes the
one we want to be default is the animated one.

Signed-off-by: Adam Williamson <awilliam@redhat.com>