Add the missing budgie XML config file
This has been in the .tar.xz tarballs Luya has uploaded since
the Budgie support was added, but it is not actually in the git
repo. I think he has it in his local checkout but forgot to git
add it, and he's been producing the tarballs from his local
checkout without using `git archive`, so the file wound up in
there despite not being in source control.

Signed-off-by: Adam Williamson <awilliam@redhat.com>